Scripts now use single_init.csh and single_end.csh
Originally, the single.csh script was used (it built a docker image).
Now that we have removed the "docker rmi" from each of the test scripts,
we need to remove the docker image to clean up.

This "docker rmi" is now handled in single_end.csh

Each of the test case scripts was regenerated to use this new logic.