Bump all machines to c5a.4xlarge
This is an attempt to see why several tests are not even attempted:
test_005, test_006, test_007, test_008, test_009
(all of these are serial, openmp, mpi). A total of 54 separate
simulations that are not undertaken.

Additionally, test_010 does not attempt openmp or mpi, and
test_018 does not try openmp.

These are weird failures where the container is built and then
moves on to the next container build.

modified:   main.tf